Replace references and raw pointers with shared
Now SimpleCCs are stored in shared pointers, and SimpleINDs have shared pointers to them as the lhs and rhs fields. This solution is crucial for the further design, because we need to store the set of all the INDs from the previous level and also the set of all the CCs that appear in any of them.